Added more features to download IMDF edits button
