+++ src/pages/conversion/places-preview-map/imdf-model-helpers/index.js
@@ -51,10 +51,50 @@ function grabToPointer(layerName, map) {
     });
 }
 
+function updateLevels(levels, selectedLevel, newValue) {
+    levels.features = levels.features.map(item => 
+        item.id === selectedLevel.id ? newValue : item
+    );
+
+    return levels;
+}
+
+function setFields(feature, selectedLevel) {
+    if (!feature.data.properties.name) { 
+        feature.data.properties.name = {}; 
+        feature.data.properties.name.en = ''; 
+    } 
+
+    if(!feature.data.properties.category) { 
+        feature.data.properties.category = 'unspecified'; 
+    } 
+
+    if(!feature.data.properties.level_id) { 
+        feature.data.properties.level_id = selectedLevel.id; 
+    } 
+
+    if(!feature.data.properties.display_point) { 
+        feature.data.properties.display_point = {}; 
+        feature.data.properties.display_point.type = 'Point'; 
+        feature.data.properties.display_point.coordinates = []; 
+    } 
+
+    if(!feature.data.properties.label) { 
+        feature.data.properties.label = ''; 
+    } 
+}
+
+function deleteUnitPrevEdits(units, selectedLevel) {
+    units.features = units.features.filter(item => item.properties.level_id !== selectedLevel.id);
+}
+
 export {
     currentEditData,
     groupAndSort,
     drawingModeChanged,
     writeToInfoPanel,
-    grabToPointer
+    grabToPointer,
+    updateLevels,
+    setFields,
+    deleteUnitPrevEdits
 };
